Extract PageView class.
